Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
reportlab==4.0.5
matplotlib==3.7.2
Pillow==10.0.0
cryptography==41.0.4